refine link.sh

+++ link.sh
@@ -1,14 +1,45 @@
 #! /bin/sh
-ln -s ~/spf13-copy-vim/.vimrc ~/.vimrc
-ln -s ~/spf13-copy-vim/.vimrc.local ~/.vimrc.local
-ln -s ~/spf13-copy-vim/.vimrc.bundles ~/.vimrc.bundles
-ln -s ~/spf13-copy-vim/.vimrc.bundles.local ~/.vimrc.bundles.local
-ln -s ~/spf13-copy-vim/.vimrc.bundles.default ~/.vimrc.bundles.default
-ln -s ~/spf13-copy-vim/.vimrc.before ~/.vimrc.before
-ln -s ~/spf13-copy-vim/.vimrc.before.local ~/.vimrc.before.local
-ln -s ~/spf13-copy-vim/.gvimrc ~/.gvimrc
-ln -s ~/spf13-copy-vim/vim ~/.vim
-ln -s ~/spf13-copy-vim/.xvimrc ~/.xvimrc
-ln -s ~/spf13-copy-vim/.zshrc ~/.zshrc
-ln -s ~/spf13-copy-vim/gitconfig/.gitconfig ~/.gitconfig
-ln -s ~/spf13-copy-vim/gitconfig/.gitmessage ~/.gitmessage
+
+
+## declare an array variable
+declare -a arr=(".vimrc"
+                ".vimrc.local"
+                ".vimrc.bundles"
+                ".vimrc.bundles.local"
+                ".vimrc.bundles.default"
+                ".vimrc.before"
+                ".vimrc.before.local"
+                ".gvimrc"
+                ".vim"
+                ".xvimrc"
+                ".zshrc")
+
+## now loop through the above array
+for i in "${arr[@]}"
+do
+    if [ -e ~/$i ];then
+        echo "file: $i exist"
+        echo "Please backup below files and rm them:"
+        echo ${arr[@]}
+        exit 0
+    fi
+done
+
+for i in "${arr[@]}"
+do
+    ln -s ~/spf13-copy-vim/$i ~/$i
+done
+
+if [ ! -e ~/.gitmessage ];then
+    ln -s ~/spf13-copy-vim/gitconfig/.gitconfig ~/.gitconfig
+else
+    echo "fail link .gitconfig; file exists"
+fi
+
+if [ ! -e ~/.gitmessage ];then
+    ln -s ~/spf13-copy-vim/gitconfig/.gitmessage ~/.gitmessage
+else 
+    echo "fail link .gitmessage; file exists"
+fi
+
+
